ITI Final Project
This project was developed as part of the ITI final assessment.
It is a recipe management system designed for a chef to manage and share his unique recipes with users.

The system includes two main parts:

Admin Dashboard (Laravel Blade):
Allows the chef or admin to add, view, edit, and delete recipes. It also includes a login system for secure access.

Public API (Laravel API):
Provides endpoints for a mobile application to display recipes, add them to favorites, view favorites, and remove them.
Users are identified by a device_id, so no authentication is required.